Splitet the features of TerminalDialog in the classes in cwm.cmd package
The new command system provides a way to register commands throw
scripts, preparing for future updates. Now the main.Main class has the
"run" method of TerminalDialog, the cmd.Command class is a binding
between syntax and Reaction, the cmd.Reaction interface is an interface
which is used to represent an action, and the cwm.CommandRegistry class
is, the name says it, a class which provides methods to add(register),
delete(deregister) and execute commands. All command-cases are wrapped
from TerminalDialog are wrapped in methods and moved to CmdMethods,
which methods are called by Reactions to prevent line noise in the
registry classes